requirements.txt
Django==2.2.6
psycopg2==2.8.4
django-cleanup==4.0.0
pytz==2019.3
setuptools==40.8.0
pillow==6.2.1
numpy==1.17.3
sqlparse==0.3.0
cmake==3.15.3
face_recognition==1.2.3
imutils==0.5.3
opencv-python==4.1.1.26
